Value function in separate file (issue #7)
