+++ agent.py
@@ -2,178 +2,9 @@ import math
 import random
 import numpy as np
 from utils import LazyStr
-# from utils import do_profile
-from discretizer import create_discretizers
 import logging
 
 logger = logging.getLogger(__name__)
-
-
-class ActionValueFunction():
-    def __init__(self, env,
-                 default_value=0,
-                 discretize_bins=12,
-                 discretize_log=False
-                 ):
-        """Initialize a Tabular Q-Value (State-Action-Value) Approximator.
-
-            Params:
-                env: the openai gym compatible environment to use
-                default_value=0: initialize state-action value (Q) table with this value
-                discretize_bins: discretize any non-discrete state dimensions with this
-                                 number of bins
-                discretize_log: use exponentially increasing instead of equal bin sizes
-                                (the further from zero, the coarser the bin size)
-        """
-        # q_table: initial q table (optional, for continuing with pre-filled approximator,
-        #          empty by default)
-
-        self.env = env
-        self.default_value = default_value
-        self.state_disc, self.action_disc = \
-            create_discretizers(self.env, discretize_bins, discretize_log)
-        self.q_table = {}
-        self.remembered_state_keys = {}
-
-    def select_action(self, observation, policy, policy_params, save=None, load=None):
-        """Get action/value tuple of selected action"""
-        if load is None:
-            discrete_state = self.state_disc.discretize(observation)
-            state_key = self._stateKeyFor(discrete_state)
-            if save is not None:
-                self.remembered_state_keys[save] = state_key
-            else:
-                state_key = self.remembered_state_keys
-        else:
-            state_key = self.remembered_state_keys[load]
-
-        action_values = self._get_action_values(state_key, self.q_table)
-        action_index, action_value = policy(action_values, policy_params)
-        action = self.action_disc.undiscretize(action_index)
-        return action, action_value
-
-    def update_value(self, observation, action, value, save=None, load=None):
-        """Update value of observation-action"""
-        if load is None:
-            discrete_state = self.state_disc.discretize(observation)
-            state_key = self._stateKeyFor(discrete_state)
-            if save is not None:
-                self.remembered_state_keys[save] = state_key
-            else:
-                state_key = self.remembered_state_keys
-        else:
-            state_key = self.remembered_state_keys[load]
-
-        discrete_action = self.action_disc.discretize(action)
-        state_action_key = state_key + '-' + str(discrete_action)
-        self.q_table[state_action_key] = value
-
-    def print_q(self, qTable):
-        for s in range(self.state_disc.space.n):
-            if self.state_disc.grid is None:
-                s = np.array(s)
-            else:  # todo
-                # indices = np.unravel_index(range(statesDisc.n), state_grid.shape)
-                s = self.state_disc.grid[s]
-            logger.info([qTable.get(self._stateKeyFor(s) + '-' + str(a), self.default_value)
-                            for a in range(self.action_disc.space.n)])
-
-    def print_q_frozenlake(self, qTable):
-        for s in range(self.state_disc.space.n):
-            s = np.array(s)
-            print([qTable.get(self._stateKeyFor(s) + "-" + str(a), 0)
-                    for a in range(self.action_disc.space.n)])
-
-
-
-    @staticmethod
-    def _stateKeyFor(discreteObs):
-        if discreteObs.shape == ():
-            d_obs = np.reshape(discreteObs, (1,))
-        else:
-            d_obs = discreteObs
-        # slow using timeit():
-        return '-'.join(map(str, d_obs))
-        # 100x faster using timeit(), but only a few percent faster in reality:
-        # return reduce(lambda a, v: str(a) + '-' + str(v), discreteObs)
-
-    def _get_action_values(self, key_start, q_table):
-        return [q_table.get(key_start + '-' + str(aIdx), self.default_value)
-                for aIdx in range(self.action_disc.space.n)]
-
-
-    # def get_state_action_values(self, observation):
-    #     discrete_state = self.state_disc.discretize(observation)
-    #     state_key = self._stateKeyFor(discrete_state)
-    #     actions = [self.action_disc.undiscretize(aIdx)
-    #                 for aIdx in range(self.action_disc.space.n)]
-    #     action_values = self._getActions(self, state_key, self.q_table)
-    #     return actions, action_values
-    #
-    # def get_state_action_value(self, observation, action):
-    #     discrete_state = self.state_disc.discretize(observation)
-    #     discrete_action = self.action_disc.discretize(action)
-    #     state_key = self._stateKeyFor(discrete_state)
-    #     state_action_key = state_key + '-' + discrete_action
-    #     return self.q_table.get(state_action_key, self.default_value)
-    #
-    # def get_best_action_and_value(self, observation):
-    #     # TODO refactor out?
-    #     # really belongs to the policy (argmax)
-    #     discrete_state = self.state_disc.discretize(observation)
-    #     state_key = self._stateKeyFor(discrete_state)
-    #     best_action_index, best_value = self._maxQ(self, state_key, self.q_table)
-    #     best_action = self.action_disc.undiscretize(best_action_index)
-    #     return best_action, best_value
-    #
-
-    # @staticmethod
-    # def _getMaxRandomTieBreak(actions):
-    #     # twice as fast as
-    #     # np.random.choice(np.flatnonzero(actions == np.max(actions)))
-    #     best = []
-    #     bestActionVal = max(actions)
-    #     for i, aval in enumerate(actions):
-    #         if aval == bestActionVal:
-    #             best.append(i)
-    #     if len(best) == 1:
-    #         return best[0]
-    #     else:
-    #         return np.random.choice(best)
-
-    # def _maxQ(self, state_key, q_table):
-    #     actions = self._getActions(state_key, q_table)
-    #     logger.debug(f"Actions for state {state_key}: %s", actions)
-    #     # bestActionIdx = np.argmax(actions)  # tie breaking: first item
-    #     # tie breaking: random item:
-    #     bestActionIdx = self._getMaxRandomTieBreak(actions)
-    #     # bestActionIdx2 = np.random.choice(np.flatnonzero(actions == np.max(actions)))
-    #     # if actions[bestActionIdx] != actions[bestActionIdx2]:
-    #     #     raise ValueError("#### ERROR IN getMaxRandomTieBreak: max values are not the same! ####")
-    #
-    #     # logger.debug(f"Best action for state {stateKey}: %s", bestActionIdx)
-    #     # testMaxQ = maxQ_old(discreteObs, qTable)
-    #     # if testMaxQ[0] != bestActionIdx or testMaxQ[1] != actions[bestActionIdx]:
-    #     #     print('WARNING: {} != {} or {} != {}'.format(testMaxQ[0], bestActionIdx, testMaxQ[1], actions[bestActionIdx]))
-    #     return bestActionIdx, actions[bestActionIdx]
-    #
-    # def _maxQ_old(self, discreteObs, qTable):
-    #     if discreteObs.shape == ():
-    #         discreteObs = np.reshape(discreteObs, (1,))
-    #     keyStart = self._stateKeyFor(discreteObs)
-    #     bestActionIdx = 0
-    #     highestVal = qTable.get(keyStart + '-' + str(bestActionIdx), self.default_value)
-    #     for aIdx in range(1, self.action_disc.space.n):  # start at 1 because we already visited 0
-    #         key = keyStart + '-' + str(aIdx)
-    #         if key in qTable:
-    #             val = qTable[key]
-    #             if val > highestVal:
-    #                 highestVal = val
-    #                 bestActionIdx = aIdx
-    #             # print('maxQ found something', key, aIdx, val)
-    #     # if highestVal == -np.inf:
-    #     #     highestVal = 0.0
-    #     return bestActionIdx, highestVal
 
 
 def greedy(list_of_state_action_values, params):
